requirements.txt
click>=6.0
colorlog
sh
requests
pyyaml
